+++ genomix/genomix-pregelix/src/main/java/edu/uci/ics/genomix/pregelix/type/Message.java
@@ -2,13 +2,13 @@ package edu.uci.ics.genomix.pregelix.type;
 
 public class Message {
 	
-	public static final int NON = 0;
-	public static final int START = 1;
-	public static final int END = 2;
+	public static final byte NON = 0;
+	public static final byte START = 1;
+	public static final byte END = 2;
 	
 	public final static class MESSAGE_CONTENT{
 		
-		public static String getContentFromCode(int code){
+		public static String getContentFromCode(byte code){
 			String r = "";
 			switch(code){
 			case NON:
